+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Feature: - `DropdownPhoneSelect`: Add `parser` prop to modify value.
+
 ## [8.1.1] - 2022-02-08
 
 Fix: