+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-Feature: - `DropdownPhoneSelect`: Add `parser` prop to modify value.
+Feature: - `DropdownPhoneSelect`: Add `normalizer` prop to modify value.
 
 ## [8.1.1] - 2022-02-08